agent/requirements.txt
aiohttp
cchardet
uvloop
urllib3

jsonrpcclient
jsonrpcserver

rdflib
pyaml

# Testing
pytest
mock
pytest-cov
coveralls

# Documentation
sphinx
sphinx-autobuild
sphinx_rtd_theme
recommonmark

aiohttp-jinja2
aiohttp-cors
bson

tensorflow<1.4.0

web3
requests

fire
hvac

feedparser
